feat: add DeleteSurgeryResponseDTO, enhance surgery creation with new request DTO, and update surgery model with performed date

+++ hospitalApi/src/main/java/com/hospitalApi/surgery/services/SurgeryServices.java
@@ -7,6 +7,8 @@ import org.springframework.stereotype.Service;
 import com.hospitalApi.consults.models.Consult;
 import com.hospitalApi.consults.port.ForConsultPort;
 import com.hospitalApi.shared.exceptions.NotFoundException;
+import com.hospitalApi.surgery.dtos.CreateSugeryRequestDTO;
+import com.hospitalApi.surgery.dtos.DeleteSurgeryResponseDTO;
 import com.hospitalApi.surgery.models.Surgery;
 import com.hospitalApi.surgery.models.SurgeryType;
 import com.hospitalApi.surgery.ports.ForSurgeryPort;
@@ -48,12 +50,15 @@ public class SurgeryServices implements ForSurgeryPort {
     }
 
     @Override
-    public boolean deleteSurgery(String surgeryId) throws NotFoundException {
+    public DeleteSurgeryResponseDTO deleteSurgery(String surgeryId) throws NotFoundException {
         if (!surgeryRepository.existsById(surgeryId)) {
             throw new NotFoundException("No se encontró la cirugía con id " + surgeryId);
         }
+        if (surgeryRepository.existsByPerformedDateIsNotNullAndId(surgeryId)) {
+            throw new IllegalStateException("No se puede eliminar la cirugía porque ya ha sido realizada.");
+        }
         surgeryRepository.deleteById(surgeryId);
-        return true;
+        return new DeleteSurgeryResponseDTO(surgeryId, "La cirugía ha sido eliminada correctamente.", true);
     }
 
     @Override
@@ -61,5 +66,22 @@ public class SurgeryServices implements ForSurgeryPort {
         forConsultPort.findById(consultId);
         List<Surgery> surgerys = surgeryRepository.findByConsultId(consultId);
         return surgerys;
+    }
+
+    @Override
+    public Surgery markSurgeryAsPerformed(String surgeryId) throws NotFoundException, IllegalStateException {
+        Surgery surgery = getSurgery(surgeryId);
+        if (surgery.getPerformedDate() != null) {
+            throw new IllegalStateException("La cirugía ya ha sido realizada.");
+        }
+        surgery.setPerformedDate(java.time.LocalDate.now());
+        surgeryRepository.save(surgery);
+        return surgery;
+    }
+
+    @Override
+    public boolean surgeryAsPerformed(String surgeryId) throws NotFoundException {
+        Surgery surgery = getSurgery(surgeryId);
+        return surgery.getPerformedDate() != null;
     }
 }
